separeted vcs.core into core.repository and core.commit; added association tables for repository and commit; updated commits fields; updated respective test cases; updated related models;
