+++ arepo/models/common/vulnerability.py
@@ -12,7 +12,6 @@ class VulnerabilityModel(Base):
     published_date = Column('published_date', DateTime, nullable=False)
     last_modified_date = Column('last_modified_date', DateTime, nullable=False)
     configurations = relationship("ConfigurationModel", backref="vulnerability")
-    commits = relationship("CommitModel", backref="vulnerability")
     source_id = Column(Integer, ForeignKey('source.id'), nullable=False)
     source = relationship("SourceModel", backref="vulnerabilities")
